requirements.txt
flask~=2.3.2
langchain==0.1.7
tiktoken==0.6.0
python-dotenv==1.0.0
openai==1.30.1